Update demos with new handling of milliseconds

+++ Dev/SourcesDemos/Game/DemoGame/Actors/Controllers/ControllerAI.h
@@ -33,7 +33,7 @@ public:
     Character* m_character;
 
     gugu::Vector2f m_direction;
-    int m_delay;
+    float m_delay;
 };
 
 }   //namespace demoproject
